project_lab_url_shortener_microservice/lab_one_url_shortener/requirements.txt
Django
djangorestframework
django-redis
python-dotenv
gunicorn
drf-spectacular
asgiref
sqlparse
tzdata
psycopg2-binary
djangorestframework-simplejwt